Warnings: Bring cata warnings to 0. Warning level X incoming! (Probably error in world <-> logon communication)

+++ src/shared/ByteBuffer.h
@@ -107,7 +107,7 @@ class SERVER_DECL ByteBuffer
 
         template <typename T> void writeBits(T value, size_t bits)
         {
-            for (int32_t i = bits - 1; i >= 0; --i)
+            for (size_t i = bits - 1; i >= 0; --i)
                 writeBit((value >> i) & 1);
         }
